+++ noodles-bam/src/async/io/reader/record_buf.rs
@@ -1,11 +1,10 @@
-use noodles_sam::{self as sam, alignment::RecordBuf};
+use noodles_sam::alignment::RecordBuf;
 use tokio::io::{self, AsyncRead};
 
 use super::read_record;
 
 pub(super) async fn read_record_buf<R>(
     reader: &mut R,
-    header: &sam::Header,
     buf: &mut Vec<u8>,
     record: &mut RecordBuf,
 ) -> io::Result<usize>
@@ -20,7 +19,7 @@ where
     };
 
     let mut src = &buf[..];
-    decode(&mut src, header, record).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
+    decode(&mut src, record).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
 
     Ok(block_size)
 }
@@ -48,10 +47,9 @@ mod tests {
         ];
 
         let mut reader = &data[..];
-        let header = sam::Header::default();
         let mut buf = Vec::new();
         let mut record = RecordBuf::default();
-        let block_size = read_record_buf(&mut reader, &header, &mut buf, &mut record).await?;
+        let block_size = read_record_buf(&mut reader, &mut buf, &mut record).await?;
 
         assert_eq!(block_size, 34);
         assert_eq!(record, RecordBuf::default());